Add noalsaerr in HautParleur
