+++ pybot/module_microphone/Microphone.py
@@ -4,26 +4,7 @@ from typing import Callable
 from speech_recognition import Microphone as SrMicrophone, Recognizer, AudioData
 from .traitement_audio import TraitementAudio
 from .utils import textual_duration_to_seconds
-
-#ALSA error handler
-from ctypes import *
-from contextlib import contextmanager
-import pyaudio
-
-ERROR_HANDLER_FUNC = CFUNCTYPE(None, c_char_p, c_int, c_char_p, c_int, c_char_p)
-
-def py_error_handler(filename, line, function, err, fmt):
-    pass
-
-c_error_handler = ERROR_HANDLER_FUNC(py_error_handler)
-
-@contextmanager
-def noalsaerr():
-    asound = cdll.LoadLibrary('libasound.so')
-    asound.snd_lib_error_set_handler(c_error_handler)
-    yield
-    asound.snd_lib_error_set_handler(None)
-
+from ..alsa_err_remover import noalsaerr
 
 _debug = logging.getLogger("Microphone").debug
 _error = logging.getLogger("Microphone").error
